Move new Person form to mixin
